+++ src/components/Molecules/Membership/Membership.md
@@ -1,5 +1,7 @@
 Membership:
 
 ```js
-<Membership />
+import data from './dev-data/data';
+
+<Membership data={data} />;
 ```
