+++ src/components/Molecules/Membership/dev-data/data.js
@@ -1,0 +1,30 @@
+const data = {
+  data: [
+    {
+      moneyBuy: {
+        id: 1,
+        value: 10,
+        description:
+          'a month could pay for two mums in the UK to attend a regular support group for postnatal depression.'
+      }
+    },
+    {
+      moneyBuy: {
+        id: 2,
+        value: 20,
+        description:
+          'a month could pay for two mums in the UK to attend a regular support group for postnatal depression.'
+      }
+    },
+    {
+      moneyBuy: {
+        id: 3,
+        value: 30,
+        description:
+          'a month could pay for two mums in the UK to attend a regular support group for postnatal depression.'
+      }
+    }
+  ]
+};
+
+export default data;
